chore: remove superfluous code Fixes #87
